requirements.txt
diffusers>=0.10.0
accelerate==1.11.0
transformers==4.57.1
ftfy==6.3.1
torch>=1.13.0
torchvision
Pillow==9.5.0
numpy
huggingface-hub==0.36.0
hf_xet
ipywidgets